+++ Changelog.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+-   Add a --controller argument
+
 ## [1.6.1] 2021-01-29
 
 -   Add MT_DEBUG_INTERVAL setting